backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1

langchain==0.3.18
langchain-openai==0.3.4
langchain-chroma==0.2.1
langgraph==0.2.74

llama-index-core==0.12.16.post1
llama-index-readers-file==0.4.5
chromadb==0.5.23
rank-bm25==0.2.2
rapidfuzz==3.11.0
tree-sitter-language-pack==0.8.0
kuzu==0.8.0
numpy==1.26.4

pytest==8.3.4
httpx==0.28.1
